Fix some error in Linux

+++ src/test/tcp_appender_test.cpp
@@ -62,7 +62,7 @@ int tcp_appender_client(std::atomic_bool &running, unsigned int log_count, unsig
 	running.store(true);
 	char buffer[1024];
 	for (unsigned int i = 0; i < log_count; i++) {
-		int len = recv(fd, buffer, sizeof(buffer) - 1, 0);
+		ssize_t len = recv(fd, buffer, sizeof(buffer) - 1, 0);
 		if (len > 0) {
 			buffer[len] = 0;
 			printf("TCP[%u]: %s", i, buffer);
